apps/worker-python/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
faster-whisper>=0.10.0
pytesseract>=0.3.10
Pillow>=10.4.0
numpy>=1.24.3
opencv-python>=4.8.1.78
yt-dlp>=2023.12.30

# Audio processing dependencies for ASR quality improvement
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.10.0

# Environment variable loading
python-dotenv>=1.0.0

# NER and Semantic Analysis dependencies
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.2
spacy>=3.7.0
scikit-learn>=1.3.0

# Test dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
pytest-mock>=3.11.0
